Update README.md
This is a python code which takes two images and matches them it matches the objects in images. It uses Oriented and Rotated Brief (ORB) to match the images.
Its user interface is developed in tkinter python. 
It makes you able to select images from your local pc and compare them. I have used this script to match laptop images to know if the laptops are same or not.